pyproject.toml
[project]
name = "kavak-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniela",email = "therealdanie@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.79.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai-agents (>=0.0.15,<0.0.16)"
]

[tool.poetry]
packages = [{include = "kavak_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
